Raman/Bragg AOM characterisation
When setting a power used for the Raman/Bragg AOMs, a percentage of max power is set. The run builders used have three methods of converting this to an instruction that the DDS will accept.

The Raman AOMS did not have a good fit to the model (P_out = P_in*sin(sqrt(x/Scale))^4 ), so I have opted to interpolate between my measured data.

It seems that Yos, Ryan, and I all have a slightly different format for our calibration data. To compensate for this, I added if statements that determine whos calibration data is being loaded so that the interpolation method works for everyone