Remote Control: Pulse Shape
When a square pulse is implemented, the mogtable.reduce method causes channel 2 to stay on.

For Gaussian pulses, the reduce method is ensures that the two channels are synchronized (e.g. you get a smooth frequency ramp). Square pulses without frequency ramps won't have this issue. So, I can run the reduce command on each channel individually (as opposed to using the synch output of one channel on the other)